- Added the FCriticalSection class in critsec.h to make it easier to   use critical sections in other places besides the OPL player.
SVN r896 (trunk)

+++ src/sound/music_mus_midiout.cpp
@@ -197,7 +197,7 @@ void MUSSong2::Precache()
 	{
 		if (used[i] < 128)
 		{
-			work[j++] = used[i];
+			work[j++] = (BYTE)used[i];
 		}
 		else if (used[i] >= 135 && used[i] <= 181)
 		{ // Percussions are 100-based, not 128-based, eh?
